% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_palette_ggp.R
\name{scale_color_tanagr}
\alias{scale_color_tanagr}
\title{Creates palette to use with ggplot2 (scale_color)}
\usage{
scale_color_tanagr(palette_name, discrete = TRUE, ...)
}
\arguments{
\item{palette_name}{the name of the palette}

\item{discrete}{whether to use a discrete or continuous color palette (boolean)}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
It generates a palette by linear interpolation for scale_color
}
\examples{
library(ggplot2)
ggplot(iris, aes(x = Sepal.Length, y = Petal.Length, color = Species)) +
 geom_point(size  = 3) +
 geom_density2d() +
 theme_bw(base_size=12) +
 scale_color_tanagr(palette_name = "stilpnia_preciosa")

ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length, color = Petal.Width)) +
 geom_point(size  = 4) +
 theme_bw(base_size=12) +
 scale_color_tanagr(palette_name = "chlorochrysa_nitidissima", discrete = FALSE)

}
