% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_palette_ggp.R
\name{scale_fill_tanagr}
\alias{scale_fill_tanagr}
\title{Creates palette to use with ggplot2 (scale_color)}
\usage{
scale_fill_tanagr(palette_name, discrete = TRUE, ...)
}
\arguments{
\item{discrete}{whether to use a discrete or continuous color palette (boolean)}

\item{...}{additional arguments to pass to scale_fill_gradientn}

\item{palette}{the name of the palette}
}
\description{
It generates a palette by linear interpolation for scale_fill
}
\examples{
library(ggplot2)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) + geom_bar() + theme(axis.text.x = element_text(angle = 45, hjust = 1), panel.background = element_blank()) + scale_fill_tanagr(palette = "stilpnia_preciosa", guide = "none")

}
