% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_palette_ggp.R
\name{tanagr_pal}
\alias{tanagr_pal}
\title{Chooses palette to use with ggplot2}
\usage{
tanagr_pal(palette_name)
}
\arguments{
\item{palette_name}{the name of the palette}
}
\description{
It generates a palette by linear interpolation. This function is not used on its own; it is called by other functions.
}
