% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_palette_br.R
\name{tanagr_palette}
\alias{tanagr_palette}
\title{Chooses palette to use with base R.}
\usage{
tanagr_palette(palette_name, n, discrete = TRUE)
}
\arguments{
\item{palette_name}{the name of the palette}

\item{n}{number of colors requested (discrete). Number of colors for interpolation (continuous).}

\item{discrete}{whether to use a discrete or continuous color palette (boolean)}
}
\description{
It generates a discrete palette with the number of colors available in the palette. If discrete = FALSE, a continuous palette is generated by linear interpolation (1:n) of the colors available in the palette.
}
\examples{
pal = tanagr_palette("bangsia_edwardsi")
data(iris)
plot(iris$Sepal.Width,
     iris$Sepal.Length,
     col = pal[as.numeric(iris$Species)],
     pch = 19,
     cex = 2)

pal = tanagr_palette("dacnis_berlepschi", n = 20, discrete = FALSE)
data(volcano)
image(volcano,
      col = pal,
      main = "Colors: Dacnis berlepschi")
#Change length of interpolation vector
pal = tanagr_palette("dacnis_berlepschi", n = 100, discrete = FALSE)
image(volcano,
      col = pal,
      main = "Colors: Dacnis berlepschi")

}
