% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{tanagr_palettes}
\alias{tanagr_palettes}
\title{Tanager-Inspired Color Palettes}
\format{An object of class \code{list} of length 16.}
\usage{
tanagr_palettes
}
\description{
A collection of color palettes inspired by the plumage of tanagers
}
\details{
All images from https://www.macaulaylibrary.org/
For each species, the catalog number in the Macaulay Library is provided
}
\examples{
names(tanagr_palettes)

}
\keyword{datasets}
