% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{visualize_palette}
\alias{visualize_palette}
\title{Visualization of palettes with birds}
\usage{
visualize_palette(palette_name)
}
\arguments{
\item{palette_name}{the name of the palette}
}
\description{
Visualize a single palette from the list of available palettes with the birds
To see list of available palettes use names(tanagr_palettes)
}
\examples{
visualize_palette("bangsia_edwardsi")
visualize_palette("tangara_velia")

}
