% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{viz_palette}
\alias{viz_palette}
\title{Visualization of palettes}
\usage{
viz_palette(palette_name)
}
\arguments{
\item{palette_name}{the name of the palette}
}
\description{
Visualize a single palette from the list of available palettes
To see list of available palettes use names(tanagr_palettes)
}
\examples{
viz_palette("bangsia_edwardsi")
viz_palette("tangara_velia")

}
